validaciones

+++ app/Http/Controllers/api/ComunaController.php
@@ -27,9 +27,21 @@ class ComunaController extends Controller
      */
     public function store(Request $request)
     {
+        
+        $validated = $request->validate([
+            'comu_nomb' => ['required',' max:255'],
+            'muni_codi' => ['required','numeric','min:1'],
+        ]);
+
+        if ($validated->fails()) {
+            return response()->json([
+                'msj' => 'Se produjo un error en la validaacion de la informacion.','statuscode' => 400
+            ]);
+        }
+
         $comuna = new Comuna();
-        $comuna->comu_nomb = $request->name;
-        $comuna->muni_codi = $request->code;
+        $comuna->comu_nomb = $request->comu_nomb;
+        $comuna->muni_codi = $request->muni_codi;
         $comuna->save();
 
         return json_encode(['comuna' => $comuna]);
@@ -42,6 +54,9 @@ class ComunaController extends Controller
     {
         
             $comuna = Comuna::find($id);
+            if (is_null($comuna)) {
+                return abort(404);
+            }
             $municipios = DB::table('tb_municipio')
                 ->orderBy('muni_nomb')
                 ->get();
@@ -54,7 +69,23 @@ class ComunaController extends Controller
      */
     public function update(Request $request, string $id)
     {
-                $comuna = Comuna::find($id);    
+        $validated = $request->validate([
+            'comu_nomb' => ['required',' max:255'],
+            'muni_codi' => ['required','numeric','min:1'],
+        ]);
+
+        if ($validated->fails()) {
+            return response()->json([
+                'msj' => 'Se produjo un error en la validaacion de la informacion.','statuscode' => 400
+            ]);
+        }
+
+
+                $comuna = Comuna::find($id); 
+                $comuna = Comuna::find($id);
+                if (is_null($comuna)) {
+                    return abort(404);
+                }   
 
                 $comuna->comu_nomb = $request->comu_nomb;
                 $comuna->muni_codi = $request->muni_codi;
@@ -70,6 +101,10 @@ class ComunaController extends Controller
     public function destroy(string $id)
     {
         $comuna = Comuna::find($id);
+        if (is_null($comuna)) {
+            return abort(404);
+        }   
+
         $comuna->delete();
     
         $comunas = DB::table('tb_comuna')
